Add a basic configuration system.
This system works around ConfigurationProviders. They feed Configuration
objects, which draw from two Providers. The first draws from a
user-specified source, while the second is the default. It tries to
"guess" as much as it can about the configuration, but it should *never*
be relied on; the user-specified Provider should do those things.