Make SimplePacketManager::PacketQueueRunnable work properly
This used to hold a lock on packetQueue during the entire execution, but
that won't work because that means nothing could ever be added to the
queue. Now, a lock is held for as little time as possible, though holding
a lock through more of the method may be needed if it turns out race
conditions/Concurrent...Exception. This might have to be removed, I'm not
sure yet.